✨ feat: create-exhibition page  into UI and model
